+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `SpaceRules`'s `check_move` and `move_entity` now take as argument a `PositionClaim` instead of an `Option<&dyn Entity<R>>`.
 - `SpaceRules`'s `move_entity` is used also to move entities out of the space model.
 - `RemoveCreature` frees the entity's position.
+- `RoundsRules`'s and `on_start` and `on_end` take as arguments the entities and the space manager objects.
 
 ## [0.2.0] - 2020-02-15
 ### Added